refactor: move slected layer dates to child
