+++ frontend/src/components/MapView/DateSelector/TimelineItems/index.test.tsx
@@ -5,7 +5,7 @@ import TimelineItems from '.';
 
 const props = {
   intersectionDates: [],
-  selectedLayerDates: [[]],
+  selectedLayers: [],
   selectedLayerTitles: [],
   dateRange: [
     {
@@ -13,6 +13,7 @@ const props = {
       label: '31 Dec 2021',
       month: 'Dec 2021',
       isFirstDay: false,
+      date: '2021-12-31',
     },
   ],
   clickDate: () => {},
@@ -20,8 +21,6 @@ const props = {
 };
 
 test('renders as expected', () => {
-  const { container } = render(
-    <TimelineItems selectedLayers={[]} {...props} />,
-  );
+  const { container } = render(<TimelineItems {...props} />);
   expect(container).toMatchSnapshot();
 });
